requirements.txt
colorama
geographiclib==1.52
gpxpy==1.4.2
haversine==2.5.1
lxml==4.6.3
numpy>=1.23.0
pandas==1.3.3
Pillow>=9.0.0
